requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==3.13.0
whitenoise==6.4.0